+++ trading_ai/orchestration/__init__.py
@@ -1,4 +1,4 @@
-from .models import PaperCycleJobCreate, PaperCycleJobView, PaperCycleRunView, PaperTradingCycleResult
+from .models import ManualPaperOrderRequest, PaperCycleJobCreate, PaperCycleJobView, PaperCycleRunView, PaperTradingCycleResult
 from .paper_trading import PaperTradingService, build_default_paper_trading_service
 from .scheduler import PaperTradingScheduler
 
@@ -7,6 +7,7 @@ __all__ = [
     "PaperCycleJobView",
     "PaperCycleRunView",
     "PaperTradingCycleResult",
+    "ManualPaperOrderRequest",
     "PaperTradingScheduler",
     "PaperTradingService",
     "build_default_paper_trading_service",